refactor(electron): one home each for metrics, pricing, scoped listing, owners and atomic writes
Extract what the new handlers had each re-implemented: electron/k8s/
metrics-source.ts (Metrics client + availability backoff + pod/node
usage), prometheus.ts (promQuery: URL, timeout, status, error wording),
pricing.ts (node info, datasets, rates — out of the 936-line cost.ts),
list-scope.ts (listScoped: cluster-wide, else namespace), owners.ts
(pod → workload), util/fs-atomic.ts (tmp+rename used by kubeconfig,
pricing and openapi). cost.ts collapses its rate ladder onto
currentRates(); rightsizing/overview/rbac/netpol/metrics consume the
shared modules; the rbac cache is context-keyed and coalesced; kcFor
short-circuits the active context; kubeconfig.ts reuses connection's
path helpers and parsed-document cache; extension-manifest moves out of
k8s/ and reports a discriminated union; optStr lives in dispatch.ts; the
new read-only commands are registered for integration tests.
