+++ electron/extensions/manifest.ts
