+++ electron/handlers/overview.ts
@@ -11,19 +11,20 @@
 // anything that still fails is reported in `partial` rather than failing the
 // whole command. The judging lives in electron/k8s/overview.ts.
 
-import {
-  Metrics,
-  type CoreV1Event,
-  type V1DaemonSet,
-  type V1Deployment,
-  type V1Job,
-  type V1Node,
-  type V1Pod,
-  type V1StatefulSet,
+import type {
+  CoreV1Event,
+  V1DaemonSet,
+  V1Deployment,
+  V1Job,
+  V1Node,
+  V1Pod,
+  V1StatefulSet,
 } from '@kubernetes/client-node';
 
-import type { HandlerCtx, HandlerMap } from '../dispatch';
-import { getAppsV1Api, getBatchV1Api, getCoreV1Api, kc } from '../k8s/client';
+import { optStr, type HandlerCtx, type HandlerMap } from '../dispatch';
+import { getAppsV1Api, getBatchV1Api, getCoreV1Api } from '../k8s/client';
+import { listScoped } from '../k8s/list-scope';
+import { nodeUsage as readNodeUsage, podUsageScoped, type NodeUsage } from '../k8s/metrics-source';
 import {
   foldPodsIntoOwners,
   nodeProblems,
@@ -34,51 +35,15 @@ import {
   summarizeNodes,
   workloadProblems,
   type ClusterOverview,
-  type NodeUsage,
   type TopPod,
 } from '../k8s/overview';
-import { parseCpu, parseMemory } from '../k8s/quantity';
 
 const WARNING_WINDOW_MS = 60 * 60_000;
 const TOP_PODS = 8;
 
-interface Listed<T> {
-  items: T[];
-  /** 'cluster' | 'namespace' | null when the list failed entirely. */
-  scope: 'cluster' | 'namespace' | null;
-}
-
-/**
- * Try the cluster-wide list; on failure (RBAC, usually) retry in `namespace`
- * when one is given. Never throws — a failed kind is reported as `scope: null`.
- */
-async function listBoth<T>(
-  clusterWide: () => Promise<{ items: T[] }>,
-  namespaced: ((ns: string) => Promise<{ items: T[] }>) | null,
-  namespace: string | null,
-): Promise<Listed<T>> {
-  try {
-    const { items } = await clusterWide();
-    return { items, scope: 'cluster' };
-  } catch {
-    if (!namespace || !namespaced) return { items: [], scope: null };
-    try {
-      const { items } = await namespaced(namespace);
-      return { items, scope: 'namespace' };
-    } catch {
-      return { items: [], scope: null };
-    }
-  }
-}
-
 async function nodeUsage(): Promise<Map<string, NodeUsage> | null> {
   try {
-    const list = await new Metrics(kc()).getNodeMetrics();
-    const map = new Map<string, NodeUsage>();
-    for (const m of list.items) {
-      map.set(m.metadata.name, { cpu: parseCpu(m.usage.cpu), memory: parseMemory(m.usage.memory) });
-    }
-    return map;
+    return await readNodeUsage();
   } catch {
     return null;
   }
@@ -86,19 +51,11 @@ async function nodeUsage(): Promise<Map<string, NodeUsage> | null> {
 
 async function topPods(namespace: string | null): Promise<{ cpu: TopPod[]; memory: TopPod[] } | null> {
   try {
-    const metrics = new Metrics(kc());
-    let list;
-    try {
-      list = await metrics.getPodMetrics();
-    } catch (err) {
-      if (!namespace) throw err;
-      list = await metrics.getPodMetrics(namespace);
-    }
-    const rows: TopPod[] = list.items.map((pm) => ({
-      name: pm.metadata.name,
-      namespace: pm.metadata.namespace,
-      cpu_usage: pm.containers.reduce((s, c) => s + parseCpu(c.usage.cpu), 0),
-      memory_usage: pm.containers.reduce((s, c) => s + parseMemory(c.usage.memory), 0),
+    const rows: TopPod[] = (await podUsageScoped(namespace)).map((p) => ({
+      name: p.name,
+      namespace: p.namespace,
+      cpu_usage: p.cpu_cores,
+      memory_usage: p.memory_bytes,
     }));
     return {
       cpu: [...rows].sort((a, b) => b.cpu_usage - a.cpu_usage).slice(0, TOP_PODS),
@@ -115,13 +72,13 @@ export async function getClusterOverview(namespace: string | null): Promise<Clus
   const batch = getBatchV1Api();
 
   const [nodes, pods, deployments, statefulsets, daemonsets, jobs, events, usage, top] = await Promise.all([
-    listBoth<V1Node>(() => core.listNode(), null, null),
-    listBoth<V1Pod>(() => core.listPodForAllNamespaces(), (ns) => core.listNamespacedPod({ namespace: ns }), namespace),
-    listBoth<V1Deployment>(() => apps.listDeploymentForAllNamespaces(), (ns) => apps.listNamespacedDeployment({ namespace: ns }), namespace),
-    listBoth<V1StatefulSet>(() => apps.listStatefulSetForAllNamespaces(), (ns) => apps.listNamespacedStatefulSet({ namespace: ns }), namespace),
-    listBoth<V1DaemonSet>(() => apps.listDaemonSetForAllNamespaces(), (ns) => apps.listNamespacedDaemonSet({ namespace: ns }), namespace),
-    listBoth<V1Job>(() => batch.listJobForAllNamespaces(), (ns) => batch.listNamespacedJob({ namespace: ns }), namespace),
-    listBoth<CoreV1Event>(
+    listScoped<V1Node>(() => core.listNode(), null, null),
+    listScoped<V1Pod>(() => core.listPodForAllNamespaces(), (ns) => core.listNamespacedPod({ namespace: ns }), namespace),
+    listScoped<V1Deployment>(() => apps.listDeploymentForAllNamespaces(), (ns) => apps.listNamespacedDeployment({ namespace: ns }), namespace),
+    listScoped<V1StatefulSet>(() => apps.listStatefulSetForAllNamespaces(), (ns) => apps.listNamespacedStatefulSet({ namespace: ns }), namespace),
+    listScoped<V1DaemonSet>(() => apps.listDaemonSetForAllNamespaces(), (ns) => apps.listNamespacedDaemonSet({ namespace: ns }), namespace),
+    listScoped<V1Job>(() => batch.listJobForAllNamespaces(), (ns) => batch.listNamespacedJob({ namespace: ns }), namespace),
+    listScoped<CoreV1Event>(
       () => core.listEventForAllNamespaces({ fieldSelector: 'type=Warning' }),
       (ns) => core.listNamespacedEvent({ namespace: ns, fieldSelector: 'type=Warning' }),
       namespace,
@@ -164,8 +121,5 @@ export async function getClusterOverview(namespace: string | null): Promise<Clus
 }
 
 export function register(handlers: HandlerMap, _ctx: HandlerCtx): void {
-  handlers.set('get_cluster_overview', async (args) => {
-    const ns = typeof args.namespace === 'string' && args.namespace.length > 0 ? args.namespace : null;
-    return getClusterOverview(ns);
-  });
+  handlers.set('get_cluster_overview', async (args) => getClusterOverview(optStr(args, 'namespace') ?? null));
 }
