+++ electron/integration/setup.ts
@@ -18,6 +18,10 @@ import * as helm from '../handlers/helm';
 import * as topology from '../handlers/topology';
 import * as security from '../handlers/security';
 import * as openapi from '../handlers/openapi';
+import * as overview from '../handlers/overview';
+import * as rightsizing from '../handlers/rightsizing';
+import * as rbac from '../handlers/rbac';
+import * as netpol from '../handlers/netpol';
 import { setActiveContext } from '../k8s/client';
 
 export const TEST_CONTEXT = process.env.KDASH_TEST_CONTEXT;
@@ -41,7 +45,7 @@ export function dispatch<T = unknown>(cmd: string, args?: Record<string, unknown
   if (!dispatcher) {
     setActiveContext(TEST_CONTEXT as string);
     const built = buildDispatcher(
-      [connection, resources, workloadOps, nodeOps, metrics, helm, topology, security, openapi],
+      [connection, resources, workloadOps, nodeOps, metrics, helm, topology, security, openapi, overview, rightsizing, rbac, netpol],
       ctx,
     );
     dispatcher = (c, a) => built.dispatch(c, a);
